Implement console.timeStamp
Summary:
# Changelog: [Internal]

Adds support for experimental non-standardized `console.timeStamp` API for capturing performance entries on a timeline. The main idea of the API is to be highly performant. More details in the corresponding RCP [1].

NOTE: Because of the `jsinspector-modern` stack gating logic, this won't be installed in production builds. This case will be solved in a separate diff, where we would introduce a polyfill for console.timeStamp.

NOTE: Fantom tests will be added on top of this diff. Yet to be validated if the `jsinspector-modern` stack runs on Fantom suites.

Differential Revision: D76284119