Fix accessibilityOrder when referencing both a parent and its children
Summary:
With the new implementation of accessibilityOrder we associate virtualView's nodes to the actual views they reperesent by calling `addChild(view)`

This however causes an issue, the view itself will populate its accessibility children which prevents them from following the axOrder if both the parent and children are mentioned in the order.

With this change we now set a delegate and a provider to every view within the view that sets an axOrder to prevent the default logic to set accessibilityChildren on views that are not considered "containers".

In the case that they are "containers" (not accessible yet mentioned in the order) We want to keep the default logic that assigns them accessibilityChildren.

Differential Revision: D76911074