Remove EventPipeConclusion and RuntimeScheduler::callExpiredTasks
Summary:
Changelog: [internal]

Before the event loop, we had some functionality in the event system to run "expired" tasks after dispatching batched events, providing a functionality similar to microtasks in the Event Loop. After the migration to the Event Loop, we no longer need this mechanism in the new architecture, so we can clean up the related code.

This diff does that.

IMPORTANT: this assumes there's nothing using `Scheduler` that's not on bridgeless, which we need to validate.

Differential Revision: D77155461