@react-native/js-polyfills: polyfill console.timeStamp
Summary:
# Changelog: [Internal]

The main reason for stubbing-out this is to make sure this method is always installed. The actual implementation will be part of the `jsinspector-modern` stack, which is fully initialized in production builds.

Once there is a gurantee that RuntimeTarget globals are always installed in any environments, we can remove polyfills altogether.

Differential Revision: D76987507