Modal "avoidKeyboard" prop (#52526)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/52526

## Changelog
[iOS] [Added] - Modal "avoidKeyboard" prop
## The problem
It seems like the default behavior is that in iOS the modal is not avoiding the keyboard, meaning that anywhere that there is a modal that can trigger the virtual keyboard, this issue can occour.
## The fix
Addign support inside the "Modal" component using a new flag called "avoidKeyboard" which is implemented by conditionally wrapping the children of the modal with "KeyboardAvoidingView" with the needed attributes to cause the modal itself to avoid the keyboard to get the desired behavior.

Reviewed By: bsalex

Differential Revision: D78010672